requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
Django==2.2.16
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==3.9.2
flake8-isort==4.0.0
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
isort==5.9.3
mccabe==0.6.1
packaging==21.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
psycopg2==2.9.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
testfixtures==6.18.0
toml==0.10.2
urllib3==1.26.6
